Dispatcher call success

+++ app/src/main/java/com/kikii/smarttsassignment/data/datasource/remote/api/dispatch/DispatchResponse.kt
@@ -80,7 +80,7 @@ data class Object(
     val unixEndTime: Long?,
 
     @SerializedName("busRound")
-    val busRound: Int,
+    val busRound: Long,
 
     @SerializedName("remainEnd")
     val remainEnd: String?,
